Update config.plist
Disable RequestBootVarRouting. This setting is actually causing some issues during reboot and shutdown. The firmware bugs and locks. As a result DEL or F12 cannot be pressed.

Disable device property injection for audio. Use boot flag for now. 28 seems to be a toss up on which output is selected on reboot. Might look into creating a layout for this board.

Few other quirk changes. See OC manual for what they do.